requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
faiss-cpu
python-dotenv
tiktoken
pypdf
sentence-transformers
